# dashboard/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Industrial Time-Series Monitoring API",
    "description": "\nREST API over a historical archive of condition-monitoring data from a metro train's\ncompressor Air Production Unit (UCI MetroPT-3, DOI 10.24432/C5VW3R).\n\n**What the data is.** 22,754,220 readings - 1,516,948 scans of 15 sensors, February to\nSeptember 2020. Coverage is 82.4%: the remaining 17.6% of the timeline has no data, in\n331 recorded gaps.\n\n**Three things this API will not do to you.**\n\n* *Silently truncate.* Any series that hits a result cap is returned with\n  `truncated: true`.\n* *Hide data quality.* 3.35% of the archive is held data from a logger freeze - non-null,\n  in range, and not a measurement. It is returned with its quality code, never filtered\n  out behind your back.\n* *Return an unusable number of points.* `resolution=auto` picks raw, hourly or daily\n  from the width of the window.\n\n**Thresholds are data, not code.** Every severity boundary traces to a row in\n`analytics.SensorBaseline` or to a setpoint published by the equipment owner.\n",
    "contact": {
      "name": "Project documentation",
      "url": "https://archive.ics.uci.edu/dataset/791/metropt+3+dataset"
    },
    "license": {
      "name": "Dataset: UCI MetroPT-3, DOI 10.24432/C5VW3R"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/anomalies": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Detected anomalies",
        "description": "Paginated anomalies, worst severity first, then most recent.\n\n``total`` is the unpaginated count, so a client can show \"23 of 1,204\" rather than\ndiscovering the end of the list by walking off it.",
        "operationId": "list_anomalies_anomalies_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "sensor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sensor Id"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Severity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "method",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ADAPTIVE_MAD, BASELINE_IQR or SETPOINT. The same instant can be flagged by more than one rule, so filtering by method answers 'which rule fired' rather than collapsing them.",
              "title": "Method"
            },
            "description": "ADAPTIVE_MAD, BASELINE_IQR or SETPOINT. The same instant can be flagged by more than one rule, so filtering by method answers 'which rule fired' rather than collapsing them."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum items to return.",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum items to return."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Items to skip, for pagination.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Items to skip, for pagination."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomalyPage"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/failures": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Documented failure events",
        "description": "The four maintenance reports published with the dataset, verbatim.\n\nEach carries its lead-up data usability. Event #1's 24-hour run-up is 69.5% held\ndata, so any statistic computed over it would describe a frozen logger rather than a\ncompressor - the API says so rather than leaving a consumer to find out.",
        "operationId": "list_failures_failures_get",
        "parameters": [
          {
            "name": "equipment_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Equipment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FailureEvent"
                  },
                  "title": "Response List Failures Failures Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Anomaly": {
        "properties": {
          "anomaly_id": {
            "type": "integer",
            "title": "Anomaly Id"
          },
          "sensor_id": {
            "type": "integer",
            "title": "Sensor Id"
          },
          "sensor_code": {
            "type": "string",
            "title": "Sensor Code"
          },
          "sensor_name": {
            "type": "string",
            "title": "Sensor Name"
          },
          "equipment_code": {
            "type": "string",
            "title": "Equipment Code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "expected_low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Low"
          },
          "expected_high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected High"
          },
          "method": {
            "type": "string",
            "title": "Method",
            "description": "Which rule fired. The same instant may be flagged by more than one."
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "operating_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OperatingState"
              },
              {
                "type": "null"
              }
            ]
          },
          "detected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Detected At"
          }
        },
        "type": "object",
        "required": [
          "anomaly_id",
          "sensor_id",
          "sensor_code",
          "sensor_name",
          "equipment_code",
          "timestamp",
          "value",
          "method",
          "score",
          "severity",
          "detected_at"
        ],
        "title": "Anomaly"
      },
      "AnomalyPage": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Anomaly"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "title": "AnomalyPage"
      },
...
      "FailureEvent": {
        "properties": {
          "failure_event_id": {
            "type": "integer",
            "title": "Failure Event Id"
          },
          "equipment_id": {
            "type": "integer",
            "title": "Equipment Id"
          },
          "equipment_code": {
            "type": "string",
            "title": "Equipment Code"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          },
          "duration_minutes": {
            "type": "integer",
            "title": "Duration Minutes"
          },
          "failure_type": {
            "type": "string",
            "title": "Failure Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference",
            "description": "The identifier used in the published failure table, verbatim - including its known defects."
          },
          "report_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Note"
          },
          "data_quality_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Quality Note"
          },
          "lead_up_24h_scans": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Up 24H Scans"
          },
          "lead_up_24h_stale_scans": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Up 24H Stale Scans"
          },
          "lead_up_24h_stale_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Up 24H Stale Pct"
          },
          "lead_up_usability": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Up Usability",
            "description": "USABLE, UNUSABLE or NO DATA. Event #1's run-up is 69.5% held data, so any statistic computed over it would describe a frozen logger."
          }
        },
        "type": "object",
        "required": [
          "failure_event_id",
          "equipment_id",
          "equipment_code",
          "start",
          "end",
          "duration_minutes",
          "failure_type",
          "severity"
        ],
        "title": "FailureEvent"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperatingState": {
        "type": "string",
        "enum": [
          "OFF",
          "OFFLOADED",
          "LOADED",
          "STARTING"
        ],
        "title": "OperatingState"
      },
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemDetail",
        "description": "RFC 7807 problem details.\n\nA machine-readable error shape, so a client can branch on ``type`` rather than\npattern-matching prose that will be reworded eventually.",
        "example": {
          "detail": "start (2020-06-01T00:00:00) must be earlier than end (2020-05-01T00:00:00).",
          "instance": "/readings/14",
          "status": 422,
          "title": "Invalid time range",
          "type": "/problems/invalid-time-range"
        }
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "NORMAL",
          "WARNING",
          "CRITICAL"
        ],
        "title": "Severity"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}